sa(8): Raise WARNS to 6 and fix warnings.

+++ usr.sbin/sa/usrdb.c
@@ -28,7 +28,6 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  * $FreeBSD: src/usr.sbin/sa/usrdb.c,v 1.8.2.2 2000/10/28 02:28:20 gallatin Exp $
- * $DragonFly: src/usr.sbin/sa/usrdb.c,v 1.4 2005/12/05 02:40:28 swildner Exp $
  */
 
 #include <sys/param.h>
@@ -37,6 +36,7 @@
 #include <err.h>
 #include <errno.h>
 #include <fcntl.h>
+#include <pwd.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -234,8 +234,8 @@ usracct_print(void)
 	while (rv == 0) {
 		memcpy(ui, data.data, sizeof(struct userinfo));
 
-		printf("%-*s %9qu ", MAXLOGNAME - 1,
-		    user_from_uid(ui->ui_uid, 0), ui->ui_calls);
+		printf("%-*s %9ju ", MAXLOGNAME - 1,
+		    user_from_uid(ui->ui_uid, 0), (uintmax_t)ui->ui_calls);
 
 		t = (double) (ui->ui_utime + ui->ui_stime) /
 		    (double) AHZ;
@@ -246,15 +246,16 @@ usracct_print(void)
 
 		/* ui->ui_calls is always != 0 */
 		if (dflag)
-			printf("%12qu%s", ui->ui_io / ui->ui_calls, "avio");
+			printf("%12ju%s", (uintmax_t)ui->ui_io / ui->ui_calls,
+			    "avio");
 		else
-			printf("%12qu%s", ui->ui_io, "tio");
+			printf("%12ju%s", (uintmax_t)ui->ui_io, "tio");
 
 		/* t is always >= 0.0001; see above */
 		if (kflag)
 			printf("%12.0f%s", ui->ui_mem / t, "k");
 		else
-			printf("%12qu%s", ui->ui_mem, "k*sec");
+			printf("%12ju%s", (uintmax_t)ui->ui_mem, "k*sec");
 
 		printf("\n");
